Research_Assistant/requirements.txt
python-dotenv
crewai
crewai[tools]
fastapi[standard]

azure-functions

azure-identity
azure.keyvault.secrets
uvicorn[standard]

